+++ R/yaml.R
@@ -5,9 +5,14 @@
 #' number of data types (see \dQuote{Examples}). In particular, it does not
 #' support values that span across multiple lines (such as multi-line text).
 #' @param x A character vector of YAML data.
-#' @param ... Arguments to be passed to [yaml::yaml.load()].
+#' @param ...,handlers Arguments to be passed to [yaml::yaml.load()].
+#' @param envir The environment in which R expressions in YAML are evaluated. To
+#'   disable the evaluation, use `envir = FALSE`.
 #' @param use_yaml Whether to use the \pkg{yaml} package.
 #' @return An R object (typically a list).
+#' @note R expressions in YAML will be returned as [expression]s when they are
+#'   not evaluated. This is different with [yaml::yaml.load()], which returns
+#'   character strings for expressions.
 #' @export
 #' @examples
 #' # test the simple parser without using the yaml package
@@ -24,12 +29,23 @@
 #' read_yaml('a:\n  - b\n  - c')
 #' # and you must use the yaml package
 #' if (loadable('yaml')) yaml_load('a:\n  - b\n  - c')
-yaml_load = function(x, ..., use_yaml = loadable('yaml')) {
-  if (use_yaml) {
-    res = try_silent(yaml::yaml.load(x, eval.expr = TRUE, ...))
-    if (!inherits(res, 'try-error')) return(res)
-    warning(paste(c(x, '\nThe above YAML metadata may be invalid:\n', res), collapse = '\n'))
-  }
+yaml_load = function(
+  x, ..., handlers = NULL, envir = parent.frame(), use_yaml = loadable('yaml')
+) {
+  if (use_yaml) return(handle_error(
+    yaml::yaml.load(x, eval.expr = FALSE, handlers = yaml_handlers(handlers, envir), ...),
+    function(e, loc) {
+      s = e$message
+      r = 'line (\\d+), column (\\d+)'
+      m = regmatches(s, regexec(r, s, perl = TRUE))[[1]]
+      if (length(m) < 3) return()
+      m = as.integer(m[-1])  # c(row, col)
+      c(
+        sprintf('Failed to parse YAML%s:', loc), '',
+        append(x, paste0(strrep(' ', m[2]), '^~~~~~'), m[1]), ''
+      )
+    }
+  ))
   # the below simple parser is quite limited
   res = list()
   r = '^( *)([^ ]+?):($|\\s+.*)'
@@ -43,7 +59,7 @@ yaml_load = function(x, ..., use_yaml = loadable('yaml')) {
   keys = NULL
   for (i in seq_along(x)) {
     keys = c(head(keys, nchar(lvl[i])/2), key[i])
-    v = if (is_blank(val[i])) list() else yaml_value(val[i])
+    v = if (is_blank(val[i])) list() else yaml_value(val[i], envir)
     # special treatment of NULL (to preserve a key with a null value)
     if (is.null(v)) {
       if (length(keys) <= 1) res[keys] = list(v) else {
@@ -56,7 +72,7 @@ yaml_load = function(x, ..., use_yaml = loadable('yaml')) {
 
 # only support logical, numeric, character values (both scalar and [] arrays),
 # and R expressions starting with !r/!expr
-yaml_value = function(x) {
+yaml_value = function(x, envir = parent.frame()) {
   v = tolower(x)
   if (v == 'null') return()
   if (grepl('^true|false$', v)) return(as.logical(x))
@@ -70,13 +86,28 @@ yaml_value = function(x) {
   }
   if (grepl('^[0-9.e+-]', v)) {
     v = suppressWarnings(as.numeric(v))
-    if (!is.na(v)) return(v)
+    if (!is.na(v)) return(if ((v2 <- as.integer(v)) == v) v2 else v)
   }
   x = gsub('^["\']|["\']$', '', x)  # remove optional quotes for strings
   if (grepl(r <- '^!(r|expr) (.+)$', x)) {
-    x = eval(parse_only(gsub(r, '\\2', x)), parent.frame())
+    x = yaml_expr(gsub(r, '\\2', x), envir)
   }
   x
+}
+
+yaml_expr = function(x, envir) {
+  x = parse_only(x)
+  if (is.environment(envir)) x = eval(x, envir)
+  x
+}
+
+# add !r and !expr handlers to support a custom eval environment, which
+# yaml::yaml.load() doesn't support by default (vubiostat/r-yaml#54)
+yaml_handlers = function(h, envir) {
+  h = as.list(h)
+  f = function(x) yaml_expr(x, envir)
+  for (i in c('r', 'expr')) if (is.null(h[[i]])) h[[i]] = f
+  h
 }
 
 #' Partition the YAML metadata and the body in a document